renaming histogrm_3d
