+++ source/consumers/Histogram3D.h
